% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw.polygons}
\alias{draw.polygons}
\title{Draw polygons on the plot window}
\usage{
draw.polygons(n, draw.reg = TRUE, col = NULL, alpha = 0.5,
  prompt.labels = TRUE, ID = NULL)
}
\arguments{
  \item{n}{Number of polygons to draw.}

  \item{draw.reg}{Whether to draw the polygons on the plot.
  TRUE/FALSE.}

  \item{col}{Color values for the polygons to be drawn.
  Ignored if \code{draw.reg=FALSE}.}

  \item{alpha}{Transparency of the colors (0..1). Ignored
  if \code{draw.reg=FALSE}.}

  \item{prompt.labels}{Whether to prompt for a label after
  each polygon is drawn (TRUE/FALSE). See Details.}

  \item{ID}{A character vector of length \code{n} of the ID
  values for each polygon}
}
\value{
An object of class \code{SpatialPolygonsDataFrame}
}
\description{
Draw polygons with the mouse and get a
SpatialPolygonsDataFrame.
}
\details{
The number of polygons to be drawn must be specified by the
argument \code{n}. If \code{prompt.labels=FALSE}, default
IDs will be constructed for each polygon.
}

