% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlayer.R
\name{gridlayer}
\alias{gridlayer}
\title{Returns a square grid covering a bounding box
If both mindim and cellsize are passed, cellsize trumps mindim}
\usage{
gridlayer(bbox, proj4string = CRS(as.character(NA)), mindim = 50,
  cellsize = NULL, plotme = FALSE, status = TRUE)
}
\description{
Returns a square grid covering a bounding box
If both mindim and cellsize are passed, cellsize trumps mindim
}

