% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hexlayer.R
\name{hexlayer}
\alias{hexlayer}
\title{Returns a hexagonal grid covering a bounding box
If both mindim and cellsize are passed, cellsize trumps mindim}
\usage{
hexlayer(bbox, proj4string = CRS(as.character(NA)), mindim = 50,
  cellsize = NULL, plotme = FALSE, status = TRUE)
}
\description{
Returns a hexagonal grid covering a bounding box
If both mindim and cellsize are passed, cellsize trumps mindim
}

