% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.gmap.R
\name{lhs.gmap}
\alias{lhs.gmap}
\title{Download background images for plots}
\usage{
lhs.gmap(lhs, id = NULL, k = NULL, r = NULL, a = NULL, s = NULL,
  hs.names = NULL, gmap = c("roadmap", "satellite", "hybrid", "terrain")[3],
  status = TRUE)
}
\arguments{
\item{lhs}{A \link[tlocoh]{LoCoH-hullset} object}

\item{id}{The name(s) of individuals to extract}

\item{k}{The k value of hullsets to extract}

\item{r}{The r value of hullsets to extract}

\item{a}{The a value of hullsets to extract}

\item{s}{The s value of hullsets to extract}

\item{hs.names}{The name(s) of saved hullsets to extract}

\item{gmap}{The type of image to download: \code{"roadmap"}, \code{"satellite"}, \code{"hybrid"}, or \code{"terrain"}}

\item{status}{Show status messages, T/F}
}
\value{
A list containing one element for each id in \code{lhs}. The list is of class \code{locoh.gmap}.
Each element is another named list containing i) a raster (in the same projection as \code{lhs}, ii) a vector
of color values, and iii) the type of image (e.g., hybrid, terrain, etc.).
}
\description{
Download background images for plots
}
\details{
This function can be used to download images for plotting hullsets. By downloading them once and
storing them as a variable, plotting is faster. One image is obtained for each id (individual).
}
\examples{
\dontrun{
require(raster)
require(dismo)
toni.bg <- lhs.gmap(toni.lhs, type="terrain")
plot(toni.lhs, iso=T, gmap=toni.bg)
}
}

