% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.iso.overlap.R
\name{lhs.iso.overlap}
\alias{lhs.iso.overlap}
\title{Compute isopleth overlap}
\usage{
lhs.iso.overlap(lhs, id = NULL, k = NULL, r = NULL, a = NULL,
  s = NULL, hs.names = NULL, iso.level = NULL, hsnames_simplify = TRUE,
  status = TRUE)
}
\arguments{
\item{lhs}{A LoCoH-hullset object}

\item{id}{The id(s) of the hullsets to compare}

\item{k}{The k value of hullsets to compare}

\item{r}{The r value of hullsets to compare}

\item{a}{The a value of hullsets to compare}

\item{s}{The s value of hullsets to compare}

\item{hs.names}{The name(s) of saved hullsets to compare}

\item{iso.level}{A numeric vector of the isopleth level(s) of interest}

\item{hsnames_simplify}{If True, will simplify the hullset names to just the IDs for the row and column names of the overlap matrices (see Return)}

\item{status}{Show messages. T/F}
}
\value{
A list object with three named elements. The \emph{spdf} contains a SpatialPolygonsDataFrame,
with a data table that saves the areas of intersection expressed in map units and proportions of the isopleth area 
for each hullset. \emph{overlap_area} contains a square matrix whose values are the area of intersection in map units.
\emph{overlap_area} contains a square matrix whose values are the areas of intersection expressed as proportions
of the each the two isopleths.
}
\description{
Computes the area of isopleth overlap
}
\details{
This function computes the area intersection of isopleths among different hullsets. 
This might be done, for example, if the hullsets belong to different individuals, and you 
want to see which individuals share space. All pairs of hullsets in \code{lhs} will be compared,
and all isopleth levels will be compared.
}
\seealso{
\code{\link[tlocoh]{isopleths}}
}
