% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lhs.iso.rast}
\alias{lhs.iso.rast}
\title{Convert isopleths to rasters}
\usage{
lhs.iso.rast(lhs, id = NULL, k = NULL, r = NULL, a = NULL, s = NULL,
  hs.names = NULL, sort.metric = NULL, iso.method = "pt.quantiles",
  raster = NULL, dimSize = 100, cell.size = NULL, sf.cell.size = 2,
  ll.round = TRUE, status = TRUE)
}
\arguments{
  \item{lhs}{A \link{LoCoH-hullset} object}

  \item{id}{The id(s) of the hullsets to create isopleths
  for}

  \item{k}{The k value of hullsets to create isopleths for}

  \item{r}{The r value of hullsets to create isopleths for}

  \item{a}{The a value of hullsets to create isopleths for}

  \item{s}{The s value of hullsets to create isopleths for}

  \item{hs.names}{The name(s) of saved hullsets to create
  isopleths for}

  \item{sort.metric}{The name(s) of hull metric(s) used to
  form isopleths that rasters should be created for}

  \item{iso.method}{The method(s) used to define isopleths
  that will be converted to raster}

  \item{raster}{A RasterLayer object that will be used to
  set the extent and cell size of the rasterized isopleth}

  \item{dimSize}{The number of cells along the largest
  dimension of the track. The according raster will be
  calculated internally. Default is 100. Ignored if
  \code{raster} is passed.}

  \item{cell.size}{The size of each square cell in map
  units. Ignored if \code{raster} is passed.}

  \item{sf.cell.size}{The number of significant figures to
  use if the cell size has be computed based on
  \code{dimSize}. Default=2. Ignored if \code{raster} is
  passed.}

  \item{ll.round}{Whether to round the lower left
  coordinate to the lowest multiple of \code{cell.size}.
  Ignored if \code{raster} is passed. (T/F)}

  \item{status}{Show status messages. T/F}
}
\value{
A LoCoH-hullset object
}
\description{
Adds rasterized version of isopleth(s) to a LoCoH-hullset
object
}
\details{
This will take exising isopleths and create raster versions
of them.
}
\examples{
\dontrun{
lhs <- lhs.iso.add(lhs)
lhs <- lhs.iso.rast(lhs)
}
}
\seealso{
\code{\link{lhs.iso.add}}
}

