% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lhs.revisit.add}
\alias{lhs.revisit.add}
\title{Add revisitation hull metrics to a LoCoH-hullset object}
\usage{
lhs.revisit.add(lhs, ta.min = NULL, ta.max = NULL, ta.cuts = NULL,
  status = TRUE)
}
\arguments{
  \item{lhs}{A LoCoH-hullset object}

  \item{ta.min}{Minimum value(s) for time away in seconds
  (numeric vector)}

  \item{ta.max}{Maximum value(s) for time away in seconds
  (numeric vector)}

  \item{ta.cuts}{A numeric vector of time values in seconds
  that define the time-away intervals}

  \item{status}{Show status messages. T/F}
}
\value{
A LoCoH-hullset object
}
\description{
Computes revisitation rate based on a minimum and maximum
time away period
}
\details{
\code{ta.min} and \code{ta.max} define the minimum and
maximum period of time (in seconds) which must pass for
another occurence in the hull to be considered a 'revisit'.
They should be the same lengths. If \code{ta.max} is NULL,
no upper bound will be set

\code{ta.cuts} is an alternative way of specifying the
time-away intervals. The time values in \code{ta.cuts} will
serve as the values between time away intervals. For
example if \code{ta.cuts = c(1000,4000,7000,10000)}, three
time-away intervals will be examined: 1000 to 4000 seconds,
4000 to 7000 seconds, and 7000 to 10000 seconds. One way to
get the values for \code{ta.cuts} is to plot the
distribution of revisit times using
\code{\link{lhs.plot.revisit}}, and then use the
\code{\link{get.vals}} function to select time values
between clusters of revisit times.
}
\seealso{
\code{\link{lhs.plot.revisit}}, \code{\link{get.vals}}
}

