% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lhs.revisit.del}
\alias{lhs.revisit.del}
\title{Delete all revisitation hull metrics in a LoCoH-hullset object}
\usage{
lhs.revisit.del(lhs, status = TRUE)
}
\arguments{
  \item{lhs}{A LoCoH-hullset object}

  \item{status}{Display summary, T/F}
}
\value{
A LoCoH-hullset object
}
\description{
Delete all revisitation hull metrics in a LoCoH-hullset
object
}
\seealso{
\code{\link{lhs.revisit.add}}
}

