% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lhs.selection.R
\name{lhs.selection}
\alias{lhs.selection}
\title{Create a hullset selection object}
\usage{
lhs.selection(lhs, k = NULL, a = NULL, r = NULL)
}
\arguments{
\item{lhs}{A LoCoH-hullset object}

\item{k}{Optional initial value for \emph{k}}

\item{a}{Optional initial value for \emph{a}}

\item{r}{Optional initial value for \emph{r}}
}
\description{
Create a hullset selection object
}
\details{
This will create a new hullset selection object of class locoh.selection. You can then feed this
object into code{\link{lhs.shiny.select}} to visually select a hullset for each individual. This is typically
done when you have create a hullset containing multiple hullsets for each individual, and you want to pick one
per individual for the rest of the analysis.
}
\seealso{
code{\link{lhs.shiny.select}}
}

