% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lhs.to.add}
\alias{lhs.to.add}
\title{Compute hull metrics for temporally overlapping hulls}
\usage{
lhs.to.add(lhs, id = "all", hs2.id = "all", maxdt = "auto",
  save.hto = TRUE, status = TRUE)
}
\arguments{
  \item{lhs}{A \link{LoCoH-hullset} object}

  \item{id}{A character vector of the hullset ids to
  compute metrics for. Can also be \code{'all'}.}

  \item{hs2.id}{A character vector of the hullset ids to
  use as the comparison hullsets. Can also be
  \code{'all'}.}

  \item{maxdt}{The maximum difference in time (in seconds)
  for two hulls to be considered 'overlapping' in time. Can
  also be \code{'auto'}, in which case half of the smallest
  of the two median sampling intervals will be used.}

  \item{save.hto}{Whether to save the list of hull indices
  that temporally overlap in the hullset, T/F}

  \item{status}{Show status messages, T/F}
}
\value{
A LoCoH-hullset object
}
\description{
Compute hull metrics for pairs of temporally overlapping
hulls of separate individuals
}
\details{
This will compute hull metrics for pairs of hulls from two
individuals. This only works for a LoCoH-hullset object
that contains hulls from two or more individuals (id's).
For each hull for individual A, for example, it will
identify the hulls in individual B that temporally overlap,
and compute the mean centroid distance (hull metric name =
to.mcd). By default \emph{to.mcd} is computed for all pairs
of individuals, but you can specify specific pairs by
passing values for \code{id} and \code{hs2.id}.

This metric is used for association analysis. You can use
this metric for example to look for spatial and temporal
patterns in how close individuals get to each other. Other
association metrics exist for spatially overlapping hulls
(see \code{\link{lhs.so.add}}).
}
\seealso{
\code{\link{lhs.so.add}}, \code{\link{lhs.merge}}
}

