% Generated by roxygen2 (4.0.2): do not edit by hand
\name{locoh.lxy}
\alias{LoCoH-xy}
\alias{locoh.lxy}
\alias{locoh.lxy-class}
\title{Class for location data}
\description{
Data class for locations with associated dates, nearest
neighbor info, parameters for a random walk null model, and
assorted meta-data
}
\details{
An object of class \code{locoh.lxy} is a list containing a
series of point locations and ancillary variables that go
with those locations (e.g., time stamps, point IDs, animal
IDs). These items are bundled together and 'ready to go' as
inputs into T-LoCoH functions. The benefits of bundling all
of the objects related to a set of point locations in a
single list object includes:

\itemize{ \item cleaning and error checking only has to be
done once when the lxy object is created (e.g., with
\code{\link{xyt.lxy}}) \item having all of the ancillary
variables together simplifies the task of passing
parameters to other functions as well as saving/retrieving
your work to disk \item the nearest neighbor lookup-table
(which can take a long time to compute) can be reused \item
locations for multiple individuals can be saved in one
object }

In code examples, objects of class \code{locoh.lxy} are
often indicated by a 'lxy' suffix, and functions that are
designed to work on LoCoH-xy objects typically start with
'lxy'.

For a complete description of the data structure, see the
vignette
\href{../doc/tlocoh_data_classes.pdf}{\emph{T-LoCoH Data
Classes}}.
}
\seealso{
\code{\link{xyt.lxy}}, \code{\link{lxy.proj.add}},
\code{\link{lxy.repair}}, \code{\link{lxy.subset}},
\code{\link{lxy.merge}}, \code{\link{lxy.exp.csv}},
\code{\link{lxy.exp.kml}}, \code{\link{lxy.exp.shp}},
\code{\link{move.lxy}}
}

