% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lxy.explore.R
\name{lxy.explore}
\alias{lxy.explore}
\title{Explore a LoCoH-xy object in a interactive map}
\usage{
lxy.explore(lxy, id = NULL, bg = c("esri_world_imagery", "none")[1],
  connect.dots = FALSE, line.popup = FALSE, status = TRUE)
}
\arguments{
  \item{lxy}{A \link{LoCoH-xy} object}

  \item{id}{The name(s) of individuals to export}

  \item{bg}{The background layer (i.e., tile) to display,
  or \code{'none'}}

  \item{connect.dots}{Whether to draw a line between
  segments, T/F}

  \item{line.popup}{Enable popup balloons on the line
  segments, T/F}

  \item{status}{Display status messages, T/F}
}
\description{
Explore a LoCoH-xy object in a zoomable clickable
interactive map
}
\details{
This function displays a Locoh-xy object in an interactive
window with a satellite image in the background. You may
zoom in and out, pan, and click on locations.

To use this function, you must be using RStudio with the
leaflet package installed. To display a satellite image in
the background, you must also be connected to the internet.
}
\examples{
\dontrun{
if (!require('devtools')) install.packages('devtools')
devtools::install_github('rstudio/leaflet')
mycon <- url("http://tlocoh.r-forge.r-project.org/toni.n5775.2005-08-22.2006-04-23.lxy.RData")
load(mycon)
close(mycon)
lxy.explore(toni.lxy, connect.dots=TRUE, line.popup=TRUE)
}
}
\seealso{
\code{\link{lxy.exp.kml}}
}

