% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lxy.sval.get.R
\name{lxy.sval.get}
\alias{lxy.sval.get}
\title{Get s values from a LoCoH-xy object}
\usage{
lxy.sval.get(lxy, ptsh, id = NULL)
}
\arguments{
\item{lxy}{A \link[tlocoh]{LoCoH-xy} object}

\item{id}{The name(s) of individuals to extract values from}

\item{The}{desired proportion of time-select hulls (0..1)}
}
\value{
A named list of \emph{s} values, with one element per id.
}
\description{
Extracts the s-values from a LoCoH-xy object for a specified proportion of time-selected hulls
}
\details{
This function extracts the \emph{s} values from the individuals in a LoCoH-xy object.
This can be useful if you want to select \emph{s} values based on a consistent proportion
of time-selected hulls.

\emph{s} values are computed by \code{\link[tlocoh]{lxy.ptsh.add}} by iteratively
trying differet values of \emph{s} for a desired proportion of time selected hulls.
This function returns the corresponding value of \emph{s} that generates the desired
ptsh within a given threshhold (see \code{\link[tlocoh]{lxy.ptsh.add}}). If no matching values
of \emph{s} were found, an empty vector will be returned for that list element. If
more than one set of ptsh tables are found (i.e, lxy.ptsh.add was run more than once),
only the first set of 's' values will be returned.
}
\seealso{
\code{\link[tlocoh]{lxy.ptsh.add}}; Vignette on locoh.lxy data class
}

