% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lxy.tumap.R
\name{lxy.tumap}
\alias{lxy.tumap}
\title{Computes the time use statistics for a LoCoH-xy object using a gridded surface}
\usage{
lxy.tumap(lxy, id = NULL, ivg = NULL, gridtype = c("hex", "square")[1],
  cellsize = NULL, mindim = 20, status = TRUE)
}
\arguments{
\item{lxy}{A LoCoH-xy object}

\item{id}{The name(s) of individuals to analyze}
}
\description{
Computes the time use statistics for a LoCoH-xy object using a gridded surface
}

