% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.locoh.tumap.R
\name{plot.locoh.tumap}
\alias{plot.locoh.tumap}
\title{Plot time use maps}
\usage{
plot.locoh.tumap(tumap, breaks = 10, nsv = TRUE, mnlv = TRUE,
  ivg = NULL, ...)
}
\description{
Plot time use maps
}

