% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.locoh.tumap.R
\name{plot.locoh.tumap}
\alias{plot.locoh.tumap}
\title{Plot time use maps}
\usage{
plot.locoh.tumap(tumap, breaks = 10, nsv = TRUE, mnlv = TRUE,
  ivg = NULL, legend = c("bottomright", "bottom", "bottomleft", "left",
  "topleft", "top", "topright", "right", "center")[7], ...)
}
\arguments{
\item{tumap}{A locoh-time use map}

\item{breaks}{The number of breaks for the symbology}

\item{nsv}{Plot the revisitation values (number of separate visits), T/F}

\item{mnlv}{Plot the proxy for visit duration (mean number of locations per visit), T/F}

\item{ivg}{The intervisit-gap interval to plot (in seconds)}

\item{legend}{Where to place the legend (set to FALSE to omit a legend)}

\item{...}{Other arguments to pass to the plot function}
}
\description{
Plot time use maps
}
\details{
If \code{cex} is passed, it will change the text size for the plot and legend
}
\seealso{
\code{\link{lxy.tumap}}
}
